♻️ refactor(store): move gateway/notification/hetero electron calls behind platform twins
Keeps `@/services/electron/*` out of the share/workbench Vite graphs by
routing gateway device info, desktop notifications, and hetero
intervention submission through .desktop.ts twins.
