+++ src/store/chat/slices/agentRun/actions/entries/localHeteroIntervention.desktop.ts
@@ -1,0 +1,1 @@
+export { heterogeneousAgentService as localHeteroAgentService } from '@/services/electron/heterogeneousAgent';
