+++ src/store/chat/slices/agentRun/actions/entries/localHeteroIntervention.ts
@@ -1,0 +1,6 @@
+import type { heterogeneousAgentService } from '@/services/electron/heterogeneousAgent';
+import { desktopOnly } from '@/utils/desktopOnly';
+
+export const localHeteroAgentService = desktopOnly<typeof heterogeneousAgentService>(
+  'heterogeneousAgentService',
+);
