+++ src/store/chat/slices/agentRun/actions/transports/gateway/localDeviceInfo.desktop.ts
@@ -1,0 +1,3 @@
+import { gatewayConnectionService } from '@/services/electron/gatewayConnection';
+
+export const getLocalDeviceInfo = async () => gatewayConnectionService.getDeviceInfo();
