+++ src/store/chat/slices/agentRun/actions/transports/gateway/localDeviceInfo.ts
@@ -1,0 +1,1 @@
+export const getLocalDeviceInfo = async (): Promise<{ deviceId?: string } | undefined> => undefined;
